# perf: faster extract_hash_tag