--- a/lib/redis_client/cluster/command.rb
+++ b/lib/redis_client/cluster/command.rb
@@ -7,7 +7,9 @@ require 'redis_client/cluster/normalized_cmd_name'
 class RedisClient
   class Cluster
     class Command
-      EMPTY_STRING = ''
+      EMPTY_STRING = "".freeze
+      LEFT_BRACKET = "{".freeze
+      RIGHT_BRACKET = "}".freeze
 
       Detail = Struct.new(
         'RedisCommand',
@@ -104,8 +106,8 @@ class RedisClient
       # @see https://redis.io/topics/cluster-spec#keys-hash-tags Keys hash tags
       def extract_hash_tag(key)
         key = key.to_s
-        s = key.index('{')
-        e = key.index('}', s.to_i + 1)
+        s = key.index(LEFT_BRACKET)
+        e = key.index(RIGHT_BRACKET, s.to_i + 1)
 
         return EMPTY_STRING if s.nil? || e.nil?